Generalize product display html generation #22
