+++ src/css/components/product.scss
@@ -1,3 +1,6 @@
+@import "../abstracts/mixins";
+@import "../abstracts/variables";
+
 .product-display {
   > *:not(:last-child) {
     margin-bottom: 8rem;
@@ -26,7 +29,7 @@
     }
   }
 
-  .four-product-group {
+  .product-group {
     display: grid;
     grid-template-columns: repeat(2, 1fr);
     grid-template-rows: repeat(2, 1fr);
