+++ src/css/main.scss
@@ -8,7 +8,7 @@
 
 @import "components/button";
 @import "components/map";
-@import "components/product";
+// @import "components/product";
 @import "components/select";
 @import "components/slider";
 @import "components/video";
